+++ libraries/AP_HAL_Empty/I2CDriver.cpp
@@ -9,15 +9,15 @@ void EmptyI2CDriver::end() {}
 void EmptyI2CDriver::setTimeout(uint16_t ms) {}
 void EmptyI2CDriver::setHighSpeed(bool active) {}
 
-uint8_t EmptyI2CDriver::write(uint8_t addr, uint8_t len, uint8_t* data)
+uint8_t EmptyI2CDriver::write(uint8_t addr, uint16_t len, uint8_t* data)
 {return 1;} 
 uint8_t EmptyI2CDriver::writeRegister(uint8_t addr, uint8_t reg, uint8_t val)
 {return 1;}
 uint8_t EmptyI2CDriver::writeRegisters(uint8_t addr, uint8_t reg,
-                               uint8_t len, uint8_t* data)
+                               uint16_t len, uint8_t* data)
 {return 1;}
 
-uint8_t EmptyI2CDriver::read(uint8_t addr, uint8_t len, uint8_t* data)
+uint8_t EmptyI2CDriver::read(uint8_t addr, uint16_t len, uint8_t* data)
 {
     memset(data, 0, len);
     return 0;
@@ -29,7 +29,7 @@ uint8_t EmptyI2CDriver::readRegister(uint8_t addr, uint8_t reg, uint8_t* data)
 }
 
 uint8_t EmptyI2CDriver::readRegisters(uint8_t addr, uint8_t reg,
-                                      uint8_t len, uint8_t* data)
+                                      uint16_t len, uint8_t* data)
 {
     memset(data, 0, len);    
     return 1;
